Update hoàn chỉnh quyền menu admin

+++ trunk/Elibrary/src/java/uit/elib/action/DeleteGroupAction.java
@@ -6,12 +6,14 @@ package uit.elib.action;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
 import uit.elib.bo.GroupBO;
 import uit.elib.bo.GroupDetailBO;
 import uit.elib.bo.UserBO;
+import uit.elib.utility.CheckGroup;
 /**
  *
  * @author HERO
@@ -32,6 +34,18 @@ public class DeleteGroupAction extends org.apache.struts.action.Action {
     public ActionForward execute(ActionMapping mapping, ActionForm form,
             HttpServletRequest request, HttpServletResponse response)
             throws Exception {
+        int checkgroup =2; //visitor 
+        HttpSession session = request.getSession();
+        if(session.getAttribute("username")!=null){ 
+            CheckGroup checkGroup = new CheckGroup();
+            checkgroup = checkGroup.Group((String)session.getAttribute("username"));
+            if(checkgroup==-1) // account has just been locked while users are accessing or  account has just expired while users are accessing
+            {   
+                 session.removeAttribute("username");
+                 session.removeAttribute("group");
+            }
+            if(checkgroup==1)
+            {          
                 String groupsID = request.getParameter("groupsID");
                 groupsID = groupsID.substring(0, groupsID.length()-1);
                 String sqlGroupDetail ="delete from `groupdetail` where GroupID in("+groupsID +")";
@@ -40,6 +54,8 @@ public class DeleteGroupAction extends org.apache.struts.action.Action {
                 UserBO.getUserBO().DeleteUser(sqlUser);
                 GroupDetailBO.getGroupDetailBO().DeleteGroupDetail(sqlGroupDetail);
                 GroupBO.getGroupBO().DeleteGroup(sqlgroup);
+            }
+        }
         return null;
     }
 }
